+++ lib/assets/javascripts/new-dashboard/pages/Data/Catalog.vue
@@ -1,5 +1,5 @@
 <template>
-  <section class="catalog-section">
+  <section class="catalog">
     <div class="container grid">
       <div class="full-width">
         <SectionTitle class="grid-cell">
@@ -12,7 +12,32 @@
         </SectionTitle>
       </div>
     </div>
-    <div class="container grid">
+
+    <div class="catalog__filter container grid" action="">
+      <div class="grid-cell grid-cell--col5 catalog__filter--dropdown">
+        <CatalogDropdown ref="dropdownCategories"
+          :title="'Choose a category'"
+          :open="true"
+          :placeholder="'Search...'"
+          :options="categories"
+          @selected="getCountries"
+          @reset="resetDatasets">
+        </CatalogDropdown>
+      </div>
+      <div class="grid-cell grid-cell--col5 catalog__filter--dropdown">
+        <CatalogDropdown ref="dropdownCountries"
+          :title="'Choose a country'"
+          :placeholder="'Choose one'"
+          :options="countries"
+          :disabled="dropdownCountriesDisabled"
+          :limitHeight="true"
+          @open="dropdownCountriesOpen"
+          @selected="getDatasets"></CatalogDropdown>
+      </div>
+    </div>
+
+    <!-- Datasets list -->
+    <div class="container grid"  v-if="datasets.length > 0">
       <div class="grid-cell grid-cell--col12">
         <CatalogListHeader :order="appliedOrder" :orderDirection="appliedOrderDirection" @changeOrder="applyOrder"></CatalogListHeader>
         <ul v-if="!isFetchingDatasets" >
@@ -38,6 +63,7 @@ import SectionTitle from 'new-dashboard/components/SectionTitle';
 import CatalogListHeader from 'new-dashboard/components/Catalog/CatalogListHeader';
 import CatalogCard from 'new-dashboard/components/Catalog/CatalogCard';
 import CatalogFakeCard from 'new-dashboard/components/Catalog/CatalogFakeCard';
+import CatalogDropdown from 'new-dashboard/components/Catalog/CatalogDropdown';
 import Pagination from 'new-dashboard/components/Pagination';
 
 export default {
@@ -47,6 +73,7 @@ export default {
     CatalogListHeader,
     CatalogCard,
     CatalogFakeCard,
+    CatalogDropdown,
     Pagination
   },
   beforeRouteUpdate (to, from, next) {
@@ -55,11 +82,19 @@ export default {
   beforeMount () {
     this.$store.dispatch('catalog/fetch');
   },
+  data () {
+    return {
+      dropdownCountriesDisabled: true,
+      dropdownCountriesOpen: false,
+    }
+  },
   computed: {
     ...mapState({
       numPages: state => state.catalog.numPages,
       currentPage: state => state.catalog.page,
       datasets: state => state.catalog.list,
+      categories: state => state.catalog.categories,
+      countries: state => state.catalog.countries,
       isFetchingDatasets: state => state.catalog.isFetching,
       numResults: state => state.catalog.numResults,
       resultsPerPage: state => state.catalog.resultsPerPage,
@@ -68,7 +103,7 @@ export default {
 
     }),
     shouldShowPagination () {
-      return !this.isFetchingDatasets && this.numResults > 0 && this.numPages > 1;
+      return !this.isFetchingDatasets && this.numPages > 1;
     },
     currentPageDatasets () {
       return this.datasets.slice(((this.currentPage - 1) * this.resultsPerPage), (this.currentPage * this.resultsPerPage));
@@ -96,6 +131,24 @@ export default {
         }
       });
       this.$store.dispatch('catalog/order', orderParams);
+    },
+    getCountries (category) {
+      this.$refs.dropdownCountries.resetInput();
+      this.dropdownCountriesDisabled = false;
+      this.$store.dispatch('catalog/fetchCountries', category)
+      window.scroll({ top: 0, left: 0 });
+    },
+    getDatasets (country) {
+      this.$store.dispatch('catalog/fetchDatasets', {
+        category: this.$refs.dropdownCategories.selected,
+        country
+      })
+      window.scroll({ top: 0, left: 0 });
+    },
+    resetDatasets () {
+      this.$refs.dropdownCountries.resetInput();
+      this.$refs.dropdownCountries.closeDropdown();
+      this.dropdownCountriesDisabled = true;
     }
   }
 };
@@ -104,8 +157,18 @@ export default {
 <style scoped lang="scss">
 @import 'new-dashboard/styles/variables';
 
-.catalog-section {
+.catalog {
+  min-height: 640px;
   margin-top: 64px;
+
+  &__filter {
+    justify-content: space-between;
+    height: 168px;
+
+    &--dropdown {
+      position: relative;
+    }
+  }
 }
 
 .full-width {